+++ app/src/main/java/io/itch/fr/quizgame/screens/EndPage.kt
@@ -11,9 +11,11 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
+import androidx.navigation.NavController
 
 @Composable
 fun EndPage(
+    navController: NavController,
     score: Int,
     onPlayAgainClicked: () -> Unit,
     onViewHistoryClicked: () -> Unit